separate tables modul
